# src/modules/channels/openapi.yaml
openapi: 3.0.3
info:
  title: Axi Connect - Channels Module API
  description: API for managing communication channels, conversations, and messages in Axi Connect
  version: 1.0.0
  contact:
    name: Axi Connect Team
    email: support@axi-connect.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.axi-connect.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /channels:
    post:
      tags:
        - Channels
      summary: Create a new communication channel
      description: Creates a new communication channel (WhatsApp, SMS, Email, etc.) with its credentials
      operationId: createChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
            example:
              name: "WhatsApp Business Principal"
              type: "WHATSAPP"
              provider: "META"
              provider_account: "573001234567"
              credentials:
                accessToken: "EAAKb8xYZ..."
                phoneNumberId: "123456789"
                appId: "987654321"
              config:
                webhookUrl: "https://miapp.com/webhooks/whatsapp"
                timeout: 5000
              company_id: 1
      responses:
        '201':
          description: Channel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Channel with this provider account already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Channels
      summary: List channels for company
      description: Retrieves a paginated list of communication channels for the authenticated company
      operationId: listChannels
      parameters:
        - name: name
          in: query
          schema:
            type: string
          description: Filter by channel name (partial match)
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/ChannelType'
          description: Filter by channel type
        - name: provider
          in: query
          schema:
            $ref: '#/components/schemas/ChannelProvider'
          description: Filter by provider
        - name: is_active
          in: query
          schema:
            type: boolean
          description: Filter by active status
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, name]
            default: created_at
          description: Sort field
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort direction
      responses:
        '200':
          description: Channels retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelsListResponse'
  /channels/{id}:
    get:
      tags:
        - Channels
      summary: Get channel by ID
      description: Retrieves detailed information about a specific communication channel
      operationId: getChannel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Channel unique identifier
      responses:
        '200':
          description: Channel retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Channels
      summary: Update channel
      description: Updates an existing communication channel's information
      operationId: updateChannel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Channel unique identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChannelRequest'
      responses:
        '200':
          description: Channel updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Channels
      summary: Delete channel (soft delete)
      description: Soft deletes a communication channel (marks as deleted)
      operationId: deleteChannel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Channel unique identifier
      responses:
        '200':
          description: Channel deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /channels/{id}/activate:
    put:
      tags:
        - Channels
      summary: Activate channel
      description: Activates a previously deactivated communication channel
      operationId: activateChannel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Channel unique identifier
      responses:
        '200':
          description: Channel activated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /channels/{id}/deactivate:
    put:
      tags:
        - Channels
      summary: Deactivate channel
      description: Deactivates a communication channel without deleting it
      operationId: deactivateChannel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Channel unique identifier
      responses:
        '200':
          description: Channel deactivated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /channels/{id}/qr:
    get:
      tags:
        - Authentication
      summary: Get QR code for channel authentication
      description: Generates or retrieves a QR code for authenticating channels that require manual login (e.g., WhatsApp Web)
      operationId: getChannelQR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Channel unique identifier
      responses:
        '200':
          description: QR code generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRCodeResponse'
        '400':
          description: Channel is already active or QR auth not supported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /channels/{id}/auth:
    post:
      tags:
        - Authentication
      summary: Complete channel authentication
      description: Completes the authentication process for a channel using a session ID
      operationId: completeChannelAuth
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Channel unique identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteAuthRequest'
      responses:
        '200':
          description: Channel authentication completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '400':
          description: Invalid session or authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # CONVERSATIONS ENDPOINTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChannelType:
      type: string
      enum: [CALL, EMAIL, WHATSAPP, TELEGRAM, INSTAGRAM, MESSENGER]
      description: Type of communication channel
    ChannelProvider:
      type: string
      enum: [META, TWILIO, CUSTOM]
      description: Provider of the communication channel
...
    CreateChannelRequest:
      type: object
      required:
        - name
        - type
        - provider
        - provider_account
        - credentials
        - company_id
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Display name for the channel
        type:
          $ref: '#/components/schemas/ChannelType'
        provider:
          $ref: '#/components/schemas/ChannelProvider'
        provider_account:
          type: string
          minLength: 1
          maxLength: 255
          description: Account identifier from the provider (phone number, email, etc.)
        credentials:
          type: object
          description: Provider-specific authentication credentials
          oneOf:
            - $ref: '#/components/schemas/MetaCredentials'
            - $ref: '#/components/schemas/TwilioCredentials'
            - $ref: '#/components/schemas/CustomCredentials'
        config:
          $ref: '#/components/schemas/ChannelConfig'
        default_agent_id:
          type: integer
          description: Default agent ID for incoming messages
        company_id:
          type: integer
          description: Company that owns this channel (must exist)
        expires_at:
          type: string
          format: date-time
          description: Optional expiration date for credentials
    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        type:
          $ref: '#/components/schemas/ChannelType'
        provider:
          $ref: '#/components/schemas/ChannelProvider'
        provider_account:
          type: string
          minLength: 1
          maxLength: 255
        config:
          $ref: '#/components/schemas/ChannelConfig'
        is_active:
          type: boolean
        default_agent_id:
          type: integer
...
    ChannelResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/ChannelType'
        config:
          $ref: '#/components/schemas/ChannelConfig'
        provider:
          $ref: '#/components/schemas/ChannelProvider'
        is_active:
          type: boolean
        provider_account:
          type: string
        default_agent_id:
          type: integer
          nullable: true
        company_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
    ChannelsListResponse:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: '#/components/schemas/ChannelResponse'
        total:
          type: integer
          description: Total number of channels
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Items skipped
...
    MetaCredentials:
      type: object
      required:
        - accessToken
        - phoneNumberId
      properties:
        accessToken:
          type: string
          description: WhatsApp Business API access token
        phoneNumberId:
          type: string
          description: WhatsApp phone number ID
        appId:
          type: string
          description: Facebook App ID
        appSecret:
          type: string
          description: Facebook App Secret
        webhookVerifyToken:
          type: string
          description: Webhook verification token
    TwilioCredentials:
      type: object
      required:
        - accountSid
        - authToken
      properties:
        accountSid:
          type: string
          description: Twilio Account SID
        authToken:
          type: string
          description: Twilio Auth Token
        phoneNumber:
          type: string
          description: Twilio phone number
    CustomCredentials:
      type: object
      description: Custom provider credentials (flexible structure)
      additionalProperties: true

    # CONFIG SCHEMAS
    ChannelConfig:
      type: object
      properties:
        webhookUrl:
          type: string
          format: uri
          description: Webhook URL for receiving messages
        phoneNumberId:
          type: string
          description: Provider-specific phone number ID
        accountSid:
          type: string
          description: Provider account SID
        apiKey:
          type: string
          description: API key for authentication
        apiSecret:
          type: string
          description: API secret for authentication
        timeout:
          type: integer
          minimum: 1000
          maximum: 30000
          description: Request timeout in milliseconds
        retries:
          type: integer
          minimum: 0
          maximum: 10
          description: Number of retry attempts
        rateLimit:
          type: object
          properties:
            requests:
              type: integer
              minimum: 1
              description: Number of requests allowed
            period:
              type: integer
              minimum: 1
              description: Time period in seconds

    # AUTHENTICATION SCHEMAS
    QRCodeResponse:
      type: object
      properties:
        qrCode:
          type: string
          description: QR code string for authentication
          example: "whatsapp-auth-uuid-1234567890"
        qrCodeUrl:
          type: string
          format: uri
          description: URL to access the QR code image
          example: "/api/channels/uuid/qr-image/uuid"
        sessionId:
          type: string
          format: uuid
          description: Authentication session ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        expiresAt:
          type: string
          format: date-time
          description: Session expiration timestamp
          example: "2024-01-15T11:00:00Z"
    CompleteAuthRequest:
      type: object
      required:
        - sessionId
      properties:
        sessionId:
          type: string
          format: uuid
          description: Authentication session ID obtained from QR endpoint
          example: "123e4567-e89b-12d3-a456-426614174000"
        metadata:
          type: object
          description: Additional authentication metadata
          example:
            phoneNumber: "+573001234567"
            authenticated: true

    # COMMON SCHEMAS
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
          nullable: true
        statusCode:
          type: integer
          example: 200
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation error"
        data:
          type: object
          nullable: true
        statusCode:
          type: integer
          example: 400